renamed root ts component
