+++ demo/src/app/pages/demo/components/advanced-example-confirm.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { LocalDataSource } from '../../../../../../ng2-smart-table';
+import { LocalDataSource } from '../../../../../../eko-point-ng2-smart-table';
 
 @Component({
   selector: 'advance-example-comfirm',
