+++ demo/src/app/pages/demo/components/basic-example-load.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { LocalDataSource } from '../../../../../../ng2-smart-table';
+import { LocalDataSource } from '../../../../../../eko-point-ng2-smart-table';
 import { BasicExampleLoadService } from './basic-example-load.service';
 
 @Component({
