+++ tsconfig.json
@@ -6,6 +6,7 @@
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "sourceMap": true,
+    "declaration": true,
     "baseUrl": "./demo/src",
     "paths": {
     },
